# shared/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Shengsheng Island Internal API
  version: 0.1.0
  description: Versioned contracts between App BFF, Agent and Pet Pipeline.
servers:
  - url: http://localhost:8000
paths:
  /healthz:
    get:
      operationId: health
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /v1/agent/turn:
    post:
      operationId: createAgentTurn
      tags: [agent]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentTurnRequest"
      responses:
        "200":
          description: Agent decision and child-facing response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentTurnResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/sessions/{session_id}/turns:
    post:
      operationId: createSessionTurn
      tags: [app]
      parameters:
        - in: path
          name: session_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentTurnRequest"
      responses:
        "200":
          description: Proxied Agent turn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentTurnResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/pets/jobs:
    post:
      operationId: createPetJob
      tags: [pet]
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetJobRequest"
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetJobResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/pets:
    post:
      operationId: createPet
      tags: [app]
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetJobRequest"
      responses:
        "202":
          description: Proxied pet job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetJobResponse"
        default:
          $ref: "#/components/responses/Error"
  /v1/pets/jobs/{job_id}:
    get:
      operationId: getPetJob
      tags: [pet]
      parameters:
        - in: path
          name: job_id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Current job state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetJobResponse"
        default:
          $ref: "#/components/responses/Error"
components:
  responses:
    Error:
      description: Standard error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthResponse:
      type: object
      required: [status, service]
      properties:
        status: { type: string, const: ok }
        service: { type: string }
    LearnerProfile:
      type: object
      required: [level, confidence, target_sentence_words]
      properties:
        level: { type: integer, minimum: 1, maximum: 5 }
        confidence: { type: number, minimum: 0, maximum: 1 }
        target_sentence_words: { type: integer, minimum: 1, maximum: 20 }
        interests:
          type: array
          items: { type: string }
        recent_topics:
          type: array
          items: { type: string }
    AudioMetadata:
      type: object
      properties:
        duration_ms: { type: integer, minimum: 0 }
        asr_confidence: { type: number, minimum: 0, maximum: 1 }
        detected_language: { type: string }
        is_human_voice: { type: boolean }
    AgentTurnRequest:
      type: object
      required: [session_id, child_id, transcript, learner_profile]
      properties:
        session_id: { type: string }
        child_id: { type: string }
        transcript: { type: string, maxLength: 1000 }
        learner_profile: { $ref: "#/components/schemas/LearnerProfile" }
        audio_metadata: { $ref: "#/components/schemas/AudioMetadata" }
        context:
          type: array
          maxItems: 10
          items:
            type: object
            required: [role, text]
            properties:
              role: { type: string, enum: [child, pet] }
              text: { type: string }
    RewardDecision:
      type: object
      required: [valid_speaking_attempt, base_voice_fruit, bonus_voice_fruit, reasons]
      properties:
        valid_speaking_attempt: { type: boolean }
        base_voice_fruit: { type: integer, minimum: 0 }
        bonus_voice_fruit: { type: integer, minimum: 0 }
        reasons:
          type: array
          items: { type: string }
    SafetyResult:
      type: object
      required: [action, reason_code]
      properties:
        action: { type: string, enum: [allow, redirect, trusted_adult] }
        reason_code: { type: string }
    AgentTurnResponse:
      type: object
      required: [turn_id, reply_text, teaching_action, reward, learner_profile, safety]
      properties:
        turn_id: { type: string }
        reply_text: { type: string }
        teaching_action: { type: string, enum: [encourage, recast, scaffold, redirect] }
        recast_text: { type: [string, "null"] }
        reward: { $ref: "#/components/schemas/RewardDecision" }
        learner_profile: { $ref: "#/components/schemas/LearnerProfile" }
        safety: { $ref: "#/components/schemas/SafetyResult" }
    PetSpecification:
      type: object
      required: [species, primary_color, personality, theme, style]
      properties:
        species: { type: string, enum: [cat, dog, fox, dinosaur, dragon] }
        primary_color: { type: string }
        personality: { type: string }
        theme: { type: string }
        style: { type: string, const: child-friendly }
        accessories:
          type: array
          maxItems: 3
          items: { type: string }
    PetJobRequest:
      type: object
      required: [child_id, description]
      properties:
        child_id: { type: string }
        description: { type: string, minLength: 1, maxLength: 500 }
        output_mode: { type: string, enum: [concept_2d, template_3d], default: concept_2d }
    PetJobResponse:
      type: object
      required: [job_id, status, created_at]
      properties:
        job_id: { type: string }
        status:
          type: string
          enum: [queued, structuring, generating_concept, review_required, generating_3d, completed, failed]
        created_at: { type: string, format: date-time }
        specification:
          anyOf:
            - $ref: "#/components/schemas/PetSpecification"
            - type: "null"
        concept_image_url: { type: [string, "null"], format: uri }
        model_url: { type: [string, "null"], format: uri }
        error_code: { type: [string, "null"] }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            retryable: { type: boolean, default: false }